Alta de alojamientos

+++ amq-backend/src/main/java/com/amq/repositories/RepositoryAlojamiento.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.amq.model.Alojamiento;
 
-public interface RepositoryAlojamiento extends JpaRepository<Alojamiento, Long> {
+public interface RepositoryAlojamiento extends JpaRepository<Alojamiento, Integer> {
 	
 }
