+++ amq-backend/src/main/java/com/amq/repositories/RepositoryUsuario.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.amq.model.Usuario;
 
-public interface RepositoryUsuario extends JpaRepository<Usuario, Long> {
+public interface RepositoryUsuario extends JpaRepository<Usuario, Integer> {
 
 }
